changes in supervisor

+++ deploy/setup.sh
@@ -14,9 +14,7 @@ locale-gen en_GB.UTF-8
 # Install Python, SQLite3, and pip
 echo "Installing dependencies..."
 apt-get update
-
-# Install necessary packages, including build-essential for compiling C extensions
-apt-get install -y python3-dev python3-venv sqlite3 python3-pip supervisor nginx git build-essential libssl-dev zlib1g-dev libpcre3 libpcre3-dev
+apt-get install -y python3-dev python3-venv sqlite3 python3-pip supervisor nginx git
 
 # Create project directory and clone the repository
 mkdir -p $PROJECT_BASE_PATH
@@ -25,17 +23,10 @@ git clone $PROJECT_GIT_URL $PROJECT_BASE_PATH
 # Set up Python virtual environment
 python3 -m venv $PROJECT_BASE_PATH/env
 
-# Activate the virtual environment
-source $PROJECT_BASE_PATH/env/bin/activate
-
-# Upgrade pip and install required Python packages inside the virtual environment
+# Install required Python packages inside the virtual environment
 $PROJECT_BASE_PATH/env/bin/pip install --upgrade pip
-$PROJECT_BASE_PATH/env/bin/pip install -r $PROJECT_BASE_PATH/requirements.txt
+$PROJECT_BASE_PATH/env/bin/pip install -r $PROJECT_BASE_PATH/requirements.txt uwsgi==2.0.21
 
-# Explicitly install uWSGI version that is compatible with your Python version
-# For Python 3.8 and above, you can use the latest version
-# Modify the version number according to your Python version
-$PROJECT_BASE_PATH/env/bin/pip install uwsgi
 # Run Django migrations
 $PROJECT_BASE_PATH/env/bin/python $PROJECT_BASE_PATH/manage.py migrate
 
@@ -47,8 +38,16 @@ supervisorctl restart coconut_api
 
 # Setup Nginx to make the coconut_api accessible
 cp $PROJECT_BASE_PATH/deploy/nginx_coconut_api.conf /etc/nginx/sites-available/coconut_api.conf
-rm /etc/nginx/sites-enabled/default
+
+# Remove the default configuration file if it exists
+if [ -f /etc/nginx/sites-enabled/default ]; then
+    rm /etc/nginx/sites-enabled/default
+fi
+
+# Link your configuration file
 ln -s /etc/nginx/sites-available/coconut_api.conf /etc/nginx/sites-enabled/coconut_api.conf
+
+# Restart Nginx service
 systemctl restart nginx.service
 
 echo "Deployment of coconut API completed successfully!"
